credit card
credit card to long

+++ carRentForCompany2/app/src/main/java/com/example/moish/carrentforcompany/model/entities/Client.java
@@ -12,7 +12,7 @@ public class Client {
     long id;
     String phoneNumber;
     String email;
-    int creditCardNumber;
+    long creditCardNumber;
 
     public Client() {
     }
